+++ test/helpers/types/special-entity-data.ts
